requirements.txt
celery==3.1.23
chardet==2.3.0
dnspython==1.12.0
Flask==0.10.1
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-WTF==0.12
itsdangerous==0.24
Jinja2==2.8
MySQL-python==1.2.3
python-nmap==0.6.0
redis==2.10.5
requests==2.9.1
simplejson==3.8.2
SQLAlchemy==1.0.12
virtualenv==15.0.1
Werkzeug==0.11.5
WTForms==2.1